# internal/http/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  chat.completionMessageRequest:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  chat.createCompletionRequest:
    properties:
      conversation_id:
        type: string
      max_tokens:
        type: integer
      messages:
        items:
          $ref: '#/definitions/chat.completionMessageRequest'
        type: array
      metadata:
        additionalProperties: {}
        type: object
      model_id:
        type: string
      stream:
        type: boolean
    type: object
  chat.createConversationRequest:
    properties:
      model_id:
        type: string
      title:
        type: string
    type: object
  chat.updateConversationRequest:
    properties:
      title:
        type: string
    required:
    - title
    type: object
  httpx.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpx.ErrorBody'
      meta: {}
      request_id:
        type: string
      success:
        type: boolean
    type: object
  httpx.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: MrChat backend API for auth, user settings, chat completions, billing,
    and admin configuration. Swagger UI is exposed at /swagger/index.html.
  title: MrChat API
  version: "0.1"
paths:
...
  /chat/completions:
    post:
      consumes:
      - application/json
      description: Main chat entry. `stream=false` returns JSON. `stream=true` returns
        `text/event-stream` SSE events with `response.start`, `response.delta`, `reasoning.delta`,
        `response.completed`, `response.failed`, and `[DONE]`.
      parameters:
      - description: Chat completion payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/chat.createCompletionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Create chat completion
      tags:
      - Chat
  /conversations:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Conversation status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: List conversations
      tags:
      - Chat
    post:
      consumes:
      - application/json
      parameters:
      - description: Conversation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/chat.createConversationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Create conversation
      tags:
      - Chat
  /conversations/{id}:
    delete:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Delete conversation
      tags:
      - Chat
    put:
      consumes:
      - application/json
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Conversation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/chat.updateConversationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: Update conversation title
      tags:
      - Chat
  /conversations/{id}/messages:
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      security:
      - BearerAuth: []
      summary: List conversation messages
      tags:
      - Chat
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter access token as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"